% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstpindia-package.R
\docType{package}
\name{fstpindia-package}
\alias{fstpindia}
\alias{fstpindia-package}
\title{fstpindia: Performance of Faecal Sludge and Co-Treatment Plants in India}
\description{
Data on 47 faecal sludge treatment plants and 22 sewage treatment plants with co-treatment of faecal sludge in eight Indian states, extracted from the 2023 Phase II evaluation of faecal sludge and co-treatment systems by the Centre for Science and Environment (CSE), New Delhi. The data cover the treatment technologies, the quality of faecal sludge, inlet and outlet samples, and the removal of organic matter, nitrogen and faecal coliforms.
}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

Authors:
\itemize{
  \item Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})
}

}
\keyword{internal}
